+++ acc_app_optimisation/gui/delayed_combo_box.py
@@ -29,12 +29,24 @@ class DelayedComboBox(QtWidgets.QComboBox):
     speed and their selection only emits a signal once their selection
     has “stabilized” – hence the signal name.
 
-    Aargs:
+    Warning:
+        The widget's delayed reaction affects even `setCurrentIndex()`
+        and `setCurrentText()`. When changing the current item through
+        these methods, any slot connected to `stableIndexChanged` or
+        `stableTextChanged` will only be called after the given timeout.
+
+        If you want to change the current combo box item and immediately
+        propagate the change through your application, use
+        `setStableIndex()` or `setStableText()` instead.
+
+    Args:
         parent: The parent widget, if any.
         timeout: The timeout in milliseconds. The stable signals are
             emitted after a selection change if no further selection
             change happens within the timeout.
     """
+
+    # pylint: disable = invalid-name
 
     stableTextChanged = QtCore.pyqtSignal(str)
     stableIndexChanged = QtCore.pyqtSignal(int)
@@ -45,23 +57,57 @@ class DelayedComboBox(QtWidgets.QComboBox):
         timeout: int = 100,
         **kwargs: t.Any,
     ) -> None:
-        super().__init__(parent, **kwargs)  # type: ignore
+        super().__init__(parent, **kwargs)
         self.currentIndexChanged.connect(self._kick_off_timer)
         self._timer = QtCore.QTimer()
         self._timer.setInterval(timeout)
         self._timer.setSingleShot(True)
         self._timer.timeout.connect(self._emit_stable_signal)
+        # Use this to block the timer from being started when
+        # setStable*() is called. In case of multi-threaded access, this
+        # is an integer instead of a boolean flag. Yes, this is
+        # paranoid.
+        self._timer_skip_count = 0
 
     def timeout(self) -> int:
         """Return the timeout in milliseconds."""
         return self._timer.interval()
 
-    def setTimeout(self, msecs: int) -> None:  # pylint: disable = invalid-name
+    def setTimeout(self, msecs: int) -> None:
         """Change the timeout in milliseconds."""
         self._timer.setInterval(msecs)
 
+    def setStableIndex(self, index: int) -> None:
+        """Set the current index and immediately stabilize.
+
+        This stops any delayed reactions that might be running in the
+        background and immediately emits the usual signals
+        (`indexChanged`, `textChanged`, `stableIndexChanged`,
+        `stableTextChanged`).
+        """
+        self._timer.stop()
+        self._timer_skip_count += 1
+        self.setCurrentIndex(index)
+        self._emit_stable_signal()
+
+    def setStableText(self, text: str) -> None:
+        """Set the current text and immediately stabilize.
+
+        This stops any delayed reactions that might be running in the
+        background and immediately emits the usual signals
+        (`indexChanged`, `textChanged`, `stableIndexChanged`,
+        `stableTextChanged`).
+        """
+        self._timer.stop()
+        self._timer_skip_count += 1
+        self.setCurrentText(text)
+        self._emit_stable_signal()
+
     def _kick_off_timer(self) -> None:
-        self._timer.start()
+        if self._timer_skip_count:
+            self._timer_skip_count -= 1
+        else:
+            self._timer.start()
 
     def _emit_stable_signal(self) -> None:
         index = self.currentIndex()